Numerous changes.
Redesign of website.
Now consists of Navigation, Content, and Footer aspects.
Navigation selections change the Content via React-Router.
Navigation is always visible at the top of the page.
Sticky footer which is always at the bottom of the page.
External links are visible at the bottom right of the window.
Contact form added to the contact page.